clang-format: PointerAlignment -> Left
We wanted to switch the pointer and reference alignment style for a
while. Given now we just had many automatic changes with the previous
style change as well as the rename, it makes sense to do this change
now.

The reasoning is mostly that this is more common in C++ in general and
that it is recommended by the CppCoreGuidelines.
https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#nl18-use-c-style-declarator-layout